requirements.txt
sentry_sdk
pydantic
uvicorn
deta
fastapi
python-dotenv
slowapi
pyOpenSSL
secure